Stable candidate
This is a stable algorithm candidate.
Fixed some issues on the algorithm side, added better error checking, removed assert macros.
Now, the key update follows a slightly different path than the ciphertext, because we use key crc only to update the key body. It helps us to hide the key better.
Unless proven otherwise by tests we're trying to apply, the algorithm and function prototypes are considered a stable candidate.