Code cleanup
No algorithmic change or header format change. 
Reflection of internal Hohha messenger code updates.
Hand-optimized functions seem to be slower than general-use functions, we disabled them by default.
We use a for loop instead of memcpy in GetRandomNumbers. It's faster, since, only a few bytes are copied each time and cost of calling memcpy is more expensive than a simple for-loop copy.
We introduce a new compile time parameter: HEADER_SALT_PADDING_SIZE
SaltProtectionPadding was a single Byte Dummy variable in packet header structure in old versions. Now, it's an array of uint8 bytes, with a static parametric size  of HEADER_SALT_PADDING_SIZE
HEADER_SALT_PADDING_SIZE is set to 1 as default in order to maintain compatibility with older versions.
In Hohha, internally, we set this parameter to 4 and we suggest you to do same