Customization sample #2
The first customization sample was just an example to demonstrate how to customize. Not a real life sample.
This is a better example for real life scenarios:

In original function, the key body is updated only by crc and salt values.
In this sample, we add a new variable in our encryptor function: LastPlainTextByte
In every jump, one byte of the key body in current jump position is XORed with last plaintext byte.