Hardened against constraint satifaction algorithm
Breaking compatibility with older versions:
I hardened the algorithm against Oscar's theoretical attacks with "constraint satisfaction" algorithm.
I still consider it theoretical, since, it supposes to have both the plaintext and salt and key body crc in order to function; which, is nearly impossible in real life.
But this is how cryptoanalysts work: In their fictif world, they suppose everything but the key must be known to the attacker. That's why, we have arguments sometimes :)
But I appreciate every critic and  try to take every precaution to every possible attack.
Thank you Oscar. 
I am always telling you that your contributions are important. We just disagree on the way you tell them.
Please keep criticizing. 
And if you write a good cryptoanalysis of the algorithm until 15 March, I promise you to give a link to it in the printed version of the Inet-tr symposium paper. 15 March is the deadline. 
And let me remember to translate this story one day :)

http://kitap.ihya.org/kissadan-hisseler/konu-1602.htm